moved CheckFolderPaths and IsSubdirectory functions from Casacore to File
