+++ src/Util/Casacore.cc
@@ -17,102 +17,6 @@
 const std::regex GILDAS_REGEX(" *[a-zA-Z]+[ .]+\\(T[a-zA-Z_]+[*.]*\\) *");
 
 /**
- * @details This function checks and resolves the given top-level and starting directories.
- * If default placeholder values ("base" or "root") are found, they are replaced accordingly.
- * The function verifies that both directories exist and are accessible, and ensures that
- * the starting directory is a valid subdirectory of the top-level directory.
- *
- * @note If `starting_string` is invalid, it is replaced with `top_level_string`.
- * @note If `starting_string` is not a subdirectory of `top_level_string`, the function logs a critical error and returns `false`.
- *
- * @warning If both `top_level_string` and `starting_string` are set to their default placeholders ("base" and "root"),
- *          the function logs a critical error and returns `false`.
- */
-bool CheckFolderPaths(std::string& top_level_string, std::string& starting_string) {
-    if (top_level_string == "base" && starting_string == "root") {
-        spdlog::critical("Must set top level or starting directory. Exiting carta.");
-
-        return false;
-    }
-
-    if (top_level_string == "base")
-
-        top_level_string = starting_string;
-
-    if (starting_string == "root")
-
-        starting_string = top_level_string;
-
-    // Check if the top-level directory exists and is accessible
-
-    fs::path top_level_path = fs::weakly_canonical(fs::absolute(fs::path(top_level_string)));
-
-    if (!fs::exists(top_level_path) || !fs::is_directory(top_level_path)) {
-        spdlog::critical("Invalid top level directory, does not exist or is not a readable directory. Exiting carta.");
-
-        return false;
-    }
-
-    top_level_string = top_level_path.string();
-
-    // Check if the starting directory exists and is accessible
-
-    fs::path starting_path = fs::weakly_canonical(fs::absolute(fs::path(starting_string)));
-
-    if (!fs::exists(starting_path) || !fs::is_directory(starting_path)) {
-        spdlog::warn("Invalid starting directory, using the provided top level directory instead.");
-
-        starting_string = top_level_string;
-
-    } else {
-        starting_string = starting_path.string();
-    }
-
-    // Check if starting directory is a subdirectory of the top-level directory
-
-    if (!IsSubdirectory(starting_path, top_level_path)) {
-        spdlog::critical("Starting {} must be a subdirectory of top level {}. Exiting carta.", starting_string, top_level_string);
-
-        return false;
-    }
-
-    return true;
-}
-
-/**
- * @details This function checks if `folder` is a subdirectory of `top_folder` by
- * resolving both paths to their absolute, weakly canonical forms and then
- * traversing up the directory hierarchy.
- *
- * @warning If `top_folder` is empty, the function will always return `true`.
- */
-bool IsSubdirectory(std::string folder, std::string top_folder) {
-    auto folder_path = fs::weakly_canonical(fs::absolute(folder));
-
-    auto parent_path = fs::weakly_canonical(fs::absolute(top_folder));
-
-    if (parent_path.empty() || folder_path == parent_path) {
-        return true;
-    }
-
-    auto current_path = folder_path;
-
-    while (current_path.has_parent_path()) {
-        current_path = current_path.parent_path();
-
-        if (current_path == parent_path) {
-            return true;
-        }
-
-        if (current_path == fs::path("/")) {
-            break;
-        }
-    }
-
-    return false;
-}
-
-/**
  * @details This function constructs an absolute file path using a specified root directory,
  * a relative subdirectory, and a file name. It checks whether the resulting file path
  * exists and is readable. If any issue is encountered, an error message is set.
