+++ test/Util/TestUtilCasacore.cc
@@ -14,168 +14,6 @@ namespace fs = std::filesystem;
 
 #include "CommonTestUtilities.h"
 
-TEST(IsSubdirectoryTest, ValidSubdirectory) {
-    EXPECT_TRUE(IsSubdirectory("/home/user/docs", "/home/user"));
-    EXPECT_TRUE(IsSubdirectory("/var/log/apache", "/var/log"));
-    EXPECT_TRUE(IsSubdirectory("/usr/local/bin", "/usr/local"));
-
-    auto pwd = TestRoot();
-    std::string top_level = (pwd / "data").string();
-    std::string starting = (pwd / "data" / "test").string();
-
-    fs::create_directories(starting);
-
-    EXPECT_TRUE(IsSubdirectory(starting, top_level));
-
-    fs::remove(starting);
-}
-
-TEST(IsSubdirectoryTest, SameDirectory) {
-    EXPECT_TRUE(IsSubdirectory("/home/user", "/home/user"));
-    EXPECT_TRUE(IsSubdirectory("/etc", "/etc"));
-}
-
-TEST(IsSubdirectoryTest, ParentDirectory) {
-    EXPECT_FALSE(IsSubdirectory("/home", "/home/user"));
-    EXPECT_FALSE(IsSubdirectory("/var", "/var/log"));
-}
-
-TEST(IsSubdirectoryTest, InvalidPaths) {
-    EXPECT_FALSE(IsSubdirectory("/this/does/not/exist", "/home/user"));
-    EXPECT_FALSE(IsSubdirectory("/home/user", "/this/does/not/exist"));
-}
-
-TEST(IsSubdirectoryTest, SubdirectoryAbs) {
-    auto pwd = TestRoot();
-    EXPECT_TRUE(IsSubdirectory((pwd / "data").string(), pwd.string()));
-    EXPECT_FALSE(IsSubdirectory(pwd.string(), (pwd / "data").string()));
-    EXPECT_TRUE(IsSubdirectory((pwd / "data/images").string(), pwd.string()));
-    EXPECT_FALSE(IsSubdirectory(pwd.string(), (pwd / "data/images").string()));
-    EXPECT_TRUE(IsSubdirectory((pwd / "data/images").string(), (pwd / "data").string()));
-    EXPECT_FALSE(IsSubdirectory((pwd / "data").string(), (pwd / "data/images").string()));
-    EXPECT_TRUE(IsSubdirectory((pwd / "data/images/fits").string(), (pwd / "data/images").string()));
-    EXPECT_FALSE(IsSubdirectory((pwd / "data/images/fits").string(), (pwd / "data/images/hdf5").string()));
-}
-
-TEST(IsSubdirectoryTest, SubdirectoryRel) {
-    fs::current_path(TestRoot());
-    EXPECT_TRUE(IsSubdirectory("./data", "./"));
-    EXPECT_FALSE(IsSubdirectory("./", "./data"));
-    EXPECT_TRUE(IsSubdirectory("./data/images", "./"));
-    EXPECT_FALSE(IsSubdirectory("./", "./data/images"));
-    EXPECT_TRUE(IsSubdirectory("./data/images", "./data"));
-    EXPECT_FALSE(IsSubdirectory("./data", "./data/images"));
-    EXPECT_TRUE(IsSubdirectory("./data/images/fits", "./data/images"));
-    EXPECT_FALSE(IsSubdirectory("./data/images/fits", "./data/images/hdf5"));
-}
-
-TEST(IsSubdirectoryTest, SubdirectorySelf) {
-    auto pwd = TestRoot();
-    EXPECT_TRUE(IsSubdirectory("/", "/"));
-    EXPECT_TRUE(IsSubdirectory("./", "./"));
-    EXPECT_TRUE(IsSubdirectory(pwd.string(), pwd.string()));
-    EXPECT_TRUE(IsSubdirectory((pwd / ".").string(), pwd.string()));
-    EXPECT_TRUE(IsSubdirectory(pwd.string(), (pwd / ".").string()));
-}
-
-TEST(IsSubdirectoryTest, ParentNotSubdirectory) {
-    auto pwd = TestRoot();
-    EXPECT_FALSE(IsSubdirectory(pwd.parent_path().string(), pwd.string()));
-    EXPECT_FALSE(IsSubdirectory((pwd / "..").string(), pwd.string()));
-    EXPECT_FALSE(IsSubdirectory("../", "./"));
-}
-
-TEST(IsSubdirectoryTest, TopIsRoot) {
-    auto pwd = TestRoot();
-    EXPECT_TRUE(IsSubdirectory(pwd.string(), "/"));
-    EXPECT_TRUE(IsSubdirectory("./", "/"));
-}
-
-TEST(IsSubdirectoryTest, SubdirectoryCheck) {
-    EXPECT_TRUE(IsSubdirectory("/data/test", "/data"));
-    EXPECT_FALSE(IsSubdirectory("/etc", "/top"));
-    EXPECT_TRUE(IsSubdirectory("/data/test/sub", "/data/test"));
-}
-
-TEST(CheckFolderPathsTest, ValidPaths) {
-    auto pwd = TestRoot();
-    std::string top_level = (pwd / "data").string();
-    std::string starting = (pwd / "data" / "test").string();
-
-    fs::create_directories(starting);
-
-    EXPECT_TRUE(CheckFolderPaths(top_level, starting));
-
-    fs::remove(starting);
-}
-
-TEST(CheckFolderPathsTest, NonExistentStartingDirectory) {
-    auto pwd = TestRoot();
-    std::string top_level = (pwd / "data").string();
-    std::string starting = (pwd / "data" / "test").string();
-
-    EXPECT_TRUE(CheckFolderPaths(top_level, starting));
-}
-
-TEST(CheckFolderPathsTest, InvalidTopLevelDirectory) {
-    std::string top_level = "/data/nonexistent";
-    auto pwd = TestRoot();
-    std::string starting = (pwd / "data" / "test").string();
-
-    fs::create_directories(starting);
-
-    EXPECT_FALSE(CheckFolderPaths(top_level, starting));
-
-    fs::remove(starting);
-}
-
-TEST(CheckFolderPathsTest, TopLevelBase) {
-    std::string top_level = "base";
-    auto pwd = TestRoot();
-    std::string starting = (pwd / "data" / "test").string();
-
-    fs::create_directories(starting);
-
-    EXPECT_TRUE(CheckFolderPaths(top_level, starting));
-    EXPECT_EQ(top_level, starting);
-
-    fs::remove(starting);
-}
-
-TEST(CheckFolderPathsTest, StartingRoot) {
-    auto pwd = TestRoot();
-    std::string top_level = pwd;
-    std::string starting = "root";
-
-    EXPECT_TRUE(CheckFolderPaths(top_level, starting));
-    EXPECT_EQ(starting, top_level);
-}
-
-TEST(CheckFolderPathsTest, StartingNotSubdirectory) {
-    auto pwd = TestRoot();
-    std::string top_level = (pwd / "data" / "parent").string();
-    std::string starting = (pwd / "data" / "another").string();
-
-    fs::create_directories(top_level);
-    fs::create_directories(starting);
-
-    EXPECT_FALSE(CheckFolderPaths(top_level, starting));
-
-    fs::remove_all(top_level);
-    fs::remove_all(starting);
-}
-
-TEST(CheckFolderPathsTest, SamePath) {
-    auto pwd = TestRoot();
-    std::string top_level = (pwd / "data" / "test_same").string();
-    std::string starting = (pwd / "data" / "test_same").string();
-
-    fs::create_directories(top_level);
-
-    EXPECT_TRUE(CheckFolderPaths(top_level, starting));
-
-    fs::remove(top_level);
-}
 
 void CreateFile(const fs::path& path, const std::string& content = "") {
     std::ofstream file(path);
